small change
road to beta 1.0 patch

+++ comrfile.py
@@ -9,6 +9,11 @@ from os.path import abspath
 ctypes.windll.kernel32.SetConsoleTitleW("LightFile") #this is for the window title
 #takes a complete path (example: C:/Users/JohnDoe/Desktop/example.txt) and removes the file name (in this case, C:/Users/JohnDoe/Desktop)
 
+#-----|LightFile|-----
+#codename: blueberry
+#version 1.1 beta
+
+
 def getPath(s):
 
     #reverse string
@@ -77,6 +82,9 @@ def getFileNameFromPath(s):
 File_ext = ".lfc"
 chunksize = 1024
 root_path = '/'
+languages_folder = "lang"
+languageENG = "ENG"
+languagePL = "PL"
 
 #----|ARDT|----
 app_root_path = getPath(abspath(getsourcefile(lambda:0)))
@@ -84,18 +92,46 @@ os.chdir(app_root_path)
 os.chdir("config")
 config_file = open ("languages.txt", 'r')
 
+print(config_file.read())
+
 #----|ARDT config file check|----
+try:
+    if (config_file == "ENG"):
+        os.chdir(root_path)
+        os.chdir(app_root_path)
+        os.chdir(languages_folder)
+        os.chdir(languageENG)
+except FileNotFoundError:
+    print("could not find the config file!")
+    time.sleep(5)
+    exit()
 
-if (config_file == "ENG"):
-    os.chdir(root_path)
-    os.chdir(app_root_path)
-    os.chdir("lang")
+try:
+    if(config_file == "PL"):
+        os.chdir(root_path)
+        os.chdir(app_root_path)
+        os.chdir(languages_folder)
+        os.chdir(languagePL)
+except FileNotFoundError:
+    print("could not find the config file!")
+    time.sleep(5)
+    exit()
+
+#for some reason python is skipping the ARDT config file check. Too bad!
+
+testtesttest = os.listdir()
+print(testtesttest)
+
+print(app_root_path + "/" + languages_folder+ "/" + languagePL)
+
+try:
+    #----|ARDT language Read to memory... thingy|----
     line88 = open("88.txt", 'r')
     line90 = open("90.txt", 'r')
     line111 = open("111.txt", 'r')
     line114 = open("114.txt", 'r')
     line117 = open("117.txt", 'r')
-    line125 = open("125.txt", 'r')
+    line125 = open("125.txt", 'r') 
     line129 = open("129.txt", 'r')
     line140 = open("140.txt", 'r')
     line145 = open("145.txt", 'r')
@@ -103,27 +139,10 @@ if (config_file == "ENG"):
     line181 = open("181.txt", 'r')
     line194 = open("194.txt", 'r')
     line212 = open("212.txt", 'r')
-
-
-if(config_file == "PL"):
-    os.chdir(root_path)
-    os.chdir(app_root_path)
-    os.chdir("lang")
-    line88 = open("88.txt", 'r')
-    line90 = open("90.txt", 'r')
-    line111 = open("111.txt", 'r')
-    line114 = open("114.txt", 'r')
-    line117 = open("117.txt", 'r')
-    line125 = open("125.txt", 'r')
-    line129 = open("129.txt", 'r')
-    line140 = open("140.txt", 'r')
-    line145 = open("145.txt", 'r')
-    line151 = open("151.txt", 'r')
-    line181 = open("181.txt", 'r')
-    line194 = open("194.txt", 'r')
-    line212 = open("212.txt", 'r')
-
-
+except FileNotFoundError:
+    print("could not find translation files!" '\n' "error: 12")
+    time.sleep(5)
+    exit()
 
 #if no extra arguments give the standard selection
 if(len(sys.argv) == 1):
@@ -152,17 +171,23 @@ os.chdir(root_path)
 try:
     os.chdir(path)
 except FileNotFoundError:
-    print(line111.read().format(path))
+    print("Directory: {0} does not exist!".format(path))
     line111.close()
     time.sleep(10)
+    print("error: 2")
+    exit()
 except NotADirectoryError:
-    print(line114.read().format(path))
+    print("{0} is not a directory!".format(path))
     line114.close()
     time.sleep(10)
+    print("error: 3")
+    exit()
 except PermissionError:
-    print(line117.read().format(path))
+    print("You do not have permissions to change to {0}".format(path))
     line117.close()
     time.sleep(10)
+    print("error: 4")
+    exit()
 
 #read the file
 
@@ -172,6 +197,8 @@ try:
 except FileNotFoundError:
     logging.critical('This file does not exist!')
     print("This file does not exist!")
+    print("error: 1")
+    exit()
 
 #somewhere here i noticed a memory leak, Too Bad!
 
@@ -179,8 +206,10 @@ start_time = time.time()
 
 print("raw size:", sys.getsizeof(str))
 
-compressed_data = zlib.compress(str, 9)
-
+try:
+    compressed_data = zlib.compress(str, 9)
+except FutureWarning:
+    print("an unknown error accured" '\n' "error: 5")
 #change to the output location
 
 did_compress = True
@@ -203,11 +232,25 @@ if (new_compr_fn == ""):
 
 #create the file and write the                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      to it
 
-createfile = open(new_compr_fn + File_ext, 'w')
-createfile.close()
-savecomp = open(new_compr_fn + File_ext, 'wb')
-savecomp.write(compressed_data)
-savecomp.close()
+try:
+    createfile = open(new_compr_fn + File_ext, 'w')
+    createfile.close()
+except FileExistsError:
+    print("a file with the same name already exists in this directory!" '\n' "are you sure you don't want to rename the file? (Y/N)")
+    rusure = input(": ")
+    if (rusure == "yes" or rusure == 'y' or rusure == "Y" ):
+        new_compr_fn = input("enter the new file name" '\n' ": ")
+        createfile = open(new_compr_fn + File_ext, 'w')
+        createfile.close()
+        savecomp = open(new_compr_fn + File_ext, 'wb')
+        savecomp.write(compressed_data)
+        savecomp.close()
+    elif (rusure == "no" or rsure == 'n' or rusure == "N"):
+        createfile = open(new_compr_fn + File_ext, 'w')
+        createfile.close()
+        savecomp = open(new_compr_fn + File_ext, 'wb')
+        savecomp.write(compressed_data)
+        savecomp.close()
 
 did_save_compressed_file = True
